makefile: Add AIX as a supported platform
These changes use AIX-specific utilities where available, and skip some that
have no AIX-equivalent (e.g. ldconfig). All of these changes are protected by
a platform-specific define, so they do not affect functioning on Linux.

None of the default Linux functionality has been altered; it will continue to
be built as before.

A new target has been added that only runs the utility check (p11sak, pkcsconf)
for easier development-time testing. It has also been added to installcheck
so that the existing checks continue to work.

Because AIX does not have Linux's `killall` (or at least one that does the same
thing), this behaviour has also been wrapped into a variable whose value is
determined by the detected platform, passed down by autoconf. On AIX, it
resolves to a simpler shell script that recreates its behaviour, whereas on
Linux it resolves to the standard `killall` utility.

AIX does not have the luxury of the convenient library search path that Linux
enjoys with ldconfig. Any search path modification _must_ be done with changes
to `LIBPATH` envar, which is clunky and will not be accepted by customers.

So we do the next best thing possible - encode the library search paths
directly into the binary with the `-blibpath` linker flag. This ensures all
binaries can find their required libraries without needing any tweaks to the
`LIBPATH` variable at runtime.

The downside of this is that ANY third-party utility that makes use of
opencryptoki will need to use this flag during its compilation, otherwise
it will be unable to find the libraries it needs at runtime.

The libpath changes are all contained within their individual makefiles.

Signed-off-by: Akash Kulhalli <akash.kulhalli@ibm.com>