man: Configure man page only if corresponding feature is enabled
This extends from the changes made for autoconf. The man pages for pkcsicsf
and pkcscca were both being configured regardless of whether those features
were enabled during configuration. This is now fixed, so they will only be
configured if their corresponding features are enabled.

Additionally, the CLEANFILES variable is now deleted from each of the man page
makefiles, because it is already tracked by automake for the `distclean` target.

Adding them to the `clean` target causes required files to be cleaned as well,
breaking the build.

Signed-off-by: Akash Kulhalli <akash.kulhalli@ibm.com>