cca: Make token aware of underlying platform differences
The CCA host library has slightly different behaviour based on the underlying
platform. This is because AIX does not support version 1 type of DES keys,
which is identified by the workstation code. This is not the case with LoZ,
where the host library can also generate version 1 DES keys with some CCA
verbs. This influences the format in which the generated key is presented in,
so that requires additional handling. The token is now aware of this fact, and
extracts information based on the binary layout of the generated key. This is
mainly a concern for DES keys - other keys are unaffected by this problem.

Furthermore, symbol resolution has also been updated for all platforms.
Previously, the dlsym operation would not check for an extended error code with
dlerror; this error was checked only once, at the end. This behaviour has been
corrected and enforced by a macro. Now every dlsym operation is correctly
checked for any errors and reported accurately immediately after the operation,
instead of only checking for the last operation and reporting its error status.
The entire operation has been simplified with a new macro, `LDSYM_VERIFY`. This
makes it simpler to specify new symbols that need to be resolved from a certain
shared library.

CCA on AIX also does not support the concept of DOMAINs, and is therefore
locked to one domain during token configuration.

Signed-off-by: Akash Kulhalli <akash.kulhalli@ibm.com>