hsm_mk_change: Handle AIX-specific flock behaviour
AIX needs special handling for flock - locking a file in shared mode needs the
file descriptor to be opened in read mode, while locking the file in exclusive
mode requires it to be opened in write mode. Because the type of locking is
determined at runtime, both the modes need to be available, hence the `O_RDWR`
mode. Linux does not have this requirement, so it retains the existing
behaviour.

Signed-off-by: Akash Kulhalli <akash.kulhalli@ibm.com>