p11sak: Adapt behaviour on AIX
To handle missing qsort_r on AIX, some functions have been updated to
externally juggle the opaque data object that qsort_r takes as a parameter.
This is done via a static object that acts as a placeholder for this opaque
object across function calls.

Special notes:
glibc returns a special pointer value when the input malloc size is zero. This
value can be safely passed to `free`, but this address must not be dereferenced
by the application itself.

POSIX does not mandate a specific behaviour in case of a zero-size malloc call,
it is left to the implementation to handle this case. AIX returns a `NULL`
without setting `errno` in this case. POSIX-compliant applications must be able
to handle this behaviour correctly.

Not handling this causes issues on AIX, since it assumes the malloc call has
failed for optional attributes (where the value size is 0). With this fix,
the `--id/-i` attribute is no longer a mandatory flag for p11sak anymore.

Signed-off-by: Akash Kulhalli <akash.kulhalli@ibm.com>