pkcshsm_mk_change: Platform-aware patches
More repeats, nothing too interesting to call out here.

Only major change here is that there is no AIX-equivalent to ELIBACC, which is
an error code that is supposed to represent an inability of the program to load
a specific symbol from a shared library with `dlsym`. That has been replaced
with a generic EINVAL code.

Signed-off-by: Akash Kulhalli <akash.kulhalli@ibm.com>