slotd: Add AIX-specific changes to daemon polling routines
`epoll` is Linux-specific. AIX has an epoll clone, but it's called `pollset`.
The API signatures are nearly identical, and is intended to be the `epoll`
equivalent for AIX. AIX-specific code blocks have been added to account for it.

Privilege management on AIX is also significantly different than on Linux. AIX
has a different set of privileges that pkcsslotd does not need, and therefore
has all of them disabled at runtime. Due to platform restrictions, the slot
daemon runs under user `slotd`, instead of the usual `pkcsslotd` on Linux.
Later versions of AIX (7.3+) do not have this restriction, but are bottlenecked
by this restriction present on 7.2. Earlier releases of AIX do not have an
equivalent implementation of pollset and are therefore not supported.

Finally, file paths that use tmpfs have been adjusted according to the platform.
On Linux, it will unconditionally use the tmpfs-directories for the appropriate
files (pid and socket files under /run), while on AIX it will use the
prefix that was mentioned during configure time. This is because AIX mounts
/var as a normal jfs2 mount point, not tmpfs as Linux does. This means AIX does
not have any advantage with using /run directly, and is therefore better to
keep the install directory self-contained. /run is not a separate mount point on
AIX either.

Signed-off-by: Akash Kulhalli <akash.kulhalli@ibm.com>