pkcsstats: Support AIX as a platform
There is one disparity on AIX due to a missing platform feature. Statistics
for a particular boot session are collected in a shared memory. However, these
shared memory segments are not stored under `/dev/shm` as on Linux - there is
no way to iterate over all the shared memory segments that a process has open.

This poses a problem because then we need to know all the UIDs to obtain the
shared memory segment (if available), which is not realistically possible.
Therefore, on AIX the ability to show statistics for _all_ users is not
available - each user must run this utility as themselves to see their stats.

Signed-off-by: Akash Kulhalli <akash.kulhalli@ibm.com>