doc: Update COPYRIGHTS, README and coding guidelines
A few new files were added that require specific copyright notices to be
displayed (err.c and getopt_long.c). These files are specific to AIX and are
not used on any other platforms.

The readme has been updated as well, mentioning AIX-specific build instructions
and pitfalls where necessary. The CCA library dependencies have been updated
to reflect what the code now demands for AIX. The base version has been updated
from 5.x to 7.1, which effectively only leaves the 4769 card as the supported
card on AIX. The requirements for Linux are unchanged.

The coding guideline document has been updated as well to outline the required
formatting for preprocessor macros.

Signed-off-by: Akash Kulhalli <akash.kulhalli@ibm.com>