p11sak: Fix segfault with --id ""
Specifying a zero length hex string as input for the CKA_ID attribute
causes a segfault. This is because BN_hex2bn() does not allocate a
BIGNUM for a zero length input string. It will thus segfault later at
BN_bn2binpad() where the NULL BIGNUM is used.

Check for this condition and reject such a hex string with an appropriate
error message.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>