P11SAK: Add support for exporting a key or certificate as URI-PEM file
A URI-PEM file contains the PKCS#11 URI of the key or certificate, but not
the key or certificate material as such. Such a URI-PEM file can be used
with the pkcs11-provider from https://github.com/latchset/pkcs11-provider.

The URI contained in the PEM file can optionally contain the PKCS#11 user
pin via the 'pin-value' query attribute of the PKCS#11 URI, or it can contain
a reference to a file containing the PKCS#11 user pin via the 'pin-source'
query attribute. If 'pin-source' is used, then the current user pin is
written into the file specified as pin source.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>