SOFT: Try to load the oqsprovider in token init to support IBM Dilithium
In preparation to support the IBM specific CKM_IBM_DILITHIUM mechanism,
try to load the oqsprovider. The oqsprovider is an optional provider for
OpenSSL 3.x and must be separately installed and configured.
It is available at https://github.com/open-quantum-safe/oqs-provider/ .

If loading of the oqsprovider fails, then mechanism CKM_IBM_DILITHIUM is
not supported by the soft token.

This is an experimental feature, it may change in an incompatible way in
the future! It is not known how the 'oqsprovider' changes in the future,
and if it then still supports the Dilithium variants known by Opencryptoki.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>