SOFT: Add support for importing IBM Dilithium keys (experimental)
Add support for the soft token to import IBM Dilithium keys from clear
key values. The clear key values can either be supplied via CKA_VALUE,
or via the individual key component attributes.

For a public Dilithium key, the CKA_VALUE attribute may contain the BER
encoded SPKI of the public key. For a private Dilithium key, it may
contain the PKCS#8 encoded private key value.

This is an experimental feature, it may change in an incompatible way in
the future! It is not known how the 'oqsprovider' changes in the future,
and if it then still supports the Dilithium variants known by Opencryptoki.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>