SOFT: Add support for sign and verify with IBM Dilithium keys (experimental)
Add support for the soft token to sign and verify with IBM Dilithium keys.

This is an experimental feature, it may change in an incompatible way in
the future! It is not known how the 'oqsprovider' changes in the future,
and if it then still supports the Dilithium variants known by Opencryptoki.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>