P11SAK: Allow to import 'oqsprovider' format PEM files for Dilithium keys
The 'oqsprovider' uses a different format for storing Dilithium keys in
PEM files. It still uses a PKCS#8 or SPKI encoding, but it does not
encode the individual key components as separate OCTET STRINGs like it
is done by Opencryptoki. The 'oqsprovider' concatenates all key components
to each other and stores the result as one OCTET STRING.

The p11sak tool can now import such 'oqsprovider' format PEM files for
Dilithium keys when option '--oqsprovider-pem' is specified. By default
it still expects the PEM file in Opencryptoki format.

This is an experimental feature, it may change in an incompatible way in
the future! It is not known how the 'oqsprovider' changes in the future,
and if it then still supports the Dilithium variants known by Opencryptoki,
or if the PEM format changes.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>