TESTCASES: tok2tok_transport: Skip AES-XTS keys for CCA or EP11
The CCA and EP11 tokens do not support AES-XTS key to be wrapped.
Furthermore, AES-XTS keys require protected key support, which might not
be enabled in the token config files.

AES-XTS keys as wrapping keys are already skipped, because the CKM_AES_XTS
mechanism does not have the CKF_WRAP and CKF_UNWRAP flags set in case of
the CCA or EP11 token.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>