COMMON: Allow a token to perform actions for a just unwrapped key
A token can either implement its own token_specific_key_unwrap() function
to perform the unwrap operations itself, or it can use common code to
perform the unwrap operation. In case the token performs the unwrap itself,
it can check if it supports the to be unwrapped key type and strength (e.g.
bit size, EC curve, etc), and act accordingly. However, if the token uses
common code for unwrapping, it has no chance to check the unwrapped key.
Thus it can happen that the common code unwraps a key with a key type or
strength that the token does not support.

Call the token_specific_set_attrs_for_new_object() function with mode
MODE_UNWRAPPED to give the token a chance to inspect the just unwrapped key
and take appropriate actions. This is only done when the common code is
used to perform the unwrap. If the token provides its own
token_specific_key_unwrap() function, then no additional call to the
token_specific_set_attrs_for_new_object() function is done, because the
token can check the unwrapped key in its token_specific_key_unwrap() function
already.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>