CCA: Add support for RSA-OAEP with SHA224, SHA256, and SHA512 on en-/decrypt
Since CCA version 8.1, RSA-OAEP can be used with SHA224, SHA256, and SHA512
for encrypt and decrypt (verbs CSNDPKE and CSNDPKD).
Trying to use these digests with an earlier CCA version results in
CKR_MECHANISM_PARAM_INVALID.

CCA requires that the OAEP MGF and the hash algorithm in the OAEP
mechanism parameter use the same digest algorithm. This is also true
for the 3 new digest algorithms.

Unfortunately, PKCS#11 does not offer a way to query if a token supports
certain MGFs with RSA-OAEP. You can only query if mechanism CKM_RSA_PKCS_OAEP
is supported at all, but not which mechanism parameter combinations are
supported.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>