LGR support for PKEY: introduce convert_key routine for CCA
The new convert_key callback function handles re-creating protected keys from
their related secure keys. It is called after an LGR when continuing to run
on the new system. This includes locking/unlocking of related key objects and
retrying the operation if another LGR happens in the middle of the process.
It also handles the case where creating the protected key from an XTS key1
happened on the old system, but creating the protected key from XTS key2
happens on the new system after an LGR. Lastly, the verification pattern of
the firmware wrapping key on the new system is saved in cca_data.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>