CCA: Fix analyse_cca_key_token for version 0 DES key tokens
Do not attempt to modify the key token content while analyzing the CV field
at offset 32 of the key token. On little-endian platforms, this would cause
the CV field to be wrong after the analysis, and the usage of such a modified
key token will result in CCA error 8/29 "The token-validation value in an
internal key token is not valid".

Work on a copy of the CV field, instead of on the field inside the key token.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>