LGR support for PKEY: use convert_key routine for AES ECB/CBC/CMAC
During live guest relocation (LGR) protected keys will become invalid and
their usage will result in a verification pattern mismatch error. Any code
that uses a protected key must therefore be able to handle this error by
re-establishing the protected key from the key material it was derived
from.

Live guest relocation may happen at any time with the exception that
processing of a single data block by a CPACF instruction can be viewed as
an atomic operation. This means that the AES algorithm always either
processes a given full block or fails entirely to process this block.

The application might perform an "in place" encryption/decryption, so that
input and output buffers are identical and data up to the point where
partial completion occurred, is already processed. As there are always
complete blocks processed, opencryptoki must be able to retry such an
operation on a block boundary with the actual CPACF parm block and running
iv or counter.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>