COMMON: Report that Dual-Crypto operations are supported
The flags field in CK_TOKEN_INFO obtained via C_GetTokenInfo can contain
flag CKF_DUAL_CRYPTO_OPERATIONS to indicate that the token supports that
a single session with the token can perform dual simultaneous cryptographic
operations (digest and encrypt, decrypt and digest, sign and encrypt, and
decrypt and sign).

Opencryptoki supports this (by calling the corresponding single function
update functions in the right sequence and with the right input), so set
this flag.

Note: this only gets into effect when the token is freshly initialized.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>