TESTCASE: Reduce run time of RSA tests by caching RSA keys
A major part of the runtime of the RSA testcases is RSA key generation,
especially when running with the soft or ICA tokens.

While generating RSA keys with different parameters must certainly also be
tested, there is no need to generate RSA keys over and over again with the
same parameters.

Cache RSA keys so that an attempt to generate a new RSA key with the same
parameters in the same session is omitted, and a cached RSA key is used
instead (if available). Such potentially cached RSA key objects must not be
destroyed by the testcases, since this would make the cached keys unusable.
Instead, the cached keys of a session are destroyed when closing the session,
or at the very end of the testcase.

Unfortunately it is not easily possible to cache RSA keys across multiple
sessions, because a key object handle is only valid within the session it
was created, so a handle from session 1 can not be used on session 2.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>