CCA: EC secure key blob import: Set CKA_IBM_PROTKEY_EXTRACTABLE
Similar to when importing secure key blobs of AES or HMAC keys, set the
CKA_IBM_PROTKEY_EXTRACTABLE attribute according to the XPRTCPAC flag
in the key usage flags of the secure key blob.

This changes the behavior: Up to now the code checked the setting of the
user supplied CKA_IBM_PROTKEY_EXTRACTABLE attribute and verified if the
XPRTCPAC flag was also set in the key blob. If not an error was returned.

It is more logical to automatically set the CKA_IBM_PROTKEY_EXTRACTABLE
attribute to the value that is read from the imported key blob.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>