ICSF: Fix covscan findings on potential integer overflows
Fix covscan warnings on cases like 'if (a - b > 0)' where both 'a' and 'b'
are unsigned types. In case 'b' is larger than 'a', then the subtraction
result may overflow because the result is also treated as unsigned type.
Fix this by using 'if (a > b)' instead.

Note that in the changed places 'a' is always larger than or equal to 'b',
so the overflow does not happen. Still, changing the code to be less
error-prone is a good thing.

Closes: https://github.com/opencryptoki/opencryptoki/issues/782

Suggested-by: Than Ngo <than@redhat.com>
Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>