TESTCASES: tok2tok_transport: Skip 2nd C_Login if the same slot is used
When the tok2tok_transport testcase is used and both, the source slot and
the target slot are the same, then skip the C_Login of the target slot,
because it is already logged in.

PKCS#11 does not allow to login the same slot twice, this would result in
CKR_USER_ALREADY_LOGGED_IN. Logging in a session on a token changes the
global login state of that token, which affects all sessions of this
token.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>