COMMON: Add support for CKM_RSA_AES_KEY_WRAP to key/encrypt/decrypt manager
Implement the CKM_RSA_AES_KEY_WRAP mechanism by internally performing mechanism
CKM_AES_KEY_GEN for generation of the temporary wrapping key, mechanism
CKM_AES_KEY_WRAP_KWP for wrapping the to-be wrapped key, and mechanism
CKM_RSA_PKCS_OAEP for wrapping the temporary AES wrapping key.
Afterwards, destroy the temporary AES key.

The temporary AES wrapping key object is generated as a session object with
CKA_HIDDEN=TRUE, so that it can not be found by means of C_FindObject.
It exists only for a very short time, but it should never be able to be found
by a PKCS#11 application whatsoever.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>