PKCSSLOTD/API/COMMON: Allow to configure a token user group
A new optional key value pair named 'usergroup' can be specified within a
slot structure to specify the user group name that is set as the owner of
the token directory. Only users that are members of this group have access
to the token and its objects. Users that are not a member of this group
will not see the token as being available (e.g. via 'pkcsconf -t' or via
'C_GetTokenInfo').

If no token group is specified, then the token is owned by the 'pkcs11' group
like before this change.

If a token user group is specified, then the pkcsslotd checks if that group
exists, and if all members of that group are also members of the 'pkcs11'
group. This is required to be able to access the pkcsslotd socket at all.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>