PKCSTOK_ADMIN: Add 'chown' command to change the owner of a token directory
The 'chown' command can be used to change the owner of all files and
directories of a token. The owner group is set to the specified group,
or to the 'pkcs11' group if no group is explicitly specified.

For all files contained in the token directories, the owner user is left
unchanged, unless the current owner user of a file or directory is not
a member of the new owner group, then the owner user is set to 'root'
for such a file.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>