PKCSTOK_ADMIN: Add 'reset' command to reset a token to its initial state
The 'reset' command can be used to reset the token to its initial state.
This also resets the token's PINs to the factory default, and removes all
token objects.

Afterwards, the token must be freshly initialized using 'pkcsconf -I, the
SO pin must be set using 'pkcsconf -P' and the USER pin must be initialized
using 'pkcsconf -u'.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>