ICSF: Unify shm name and support multiple token instances
Unify the shared memory segment naming of the ICSF tokens to match the
naming of the other tokens.

Also support to have multiple ICSF tokens, by automatically adding a 'tokname'
keyword pair to the slot configuration when adding a ICSF token via the
pkcsicsf tool. For multiple ICSF tokens it is important that each token
uses its own token directory, and thus a unique token name is required.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>