COMMON: Fix coverity finding for missing unlock
When object_put() is called with unlock = FALSE, then it is assumed that
the object is currently not locked. Thus object_put() obtains the read lock
to check if the object is a session or private object. After that the object
must always be unlocked, regardless of the 'unlock' parameter. Either
'unlock' was TRUE then we must unlock the object, or if 'unlock' was FALSE,
then we locked the object our-self, and thus must unlock it again.

Currently, object_put() may only be called with unlock = FALSE by
h_opaque_2_blob() from the EP11 token, when h_opaque_2_blob() was called
with lock_type=NO_LOCK. Currently there is no such case that does that.

Another case where object_put() may be called with unlock = FALSE is
object_mgr_find_in_map_nocache(), when it gets called with lock_type=NO_LOCK.
There, in error cases object_put() may be called with unlock = FALSE.
Currently there is no such case that calls object_mgr_find_in_map_nocache()
with lock_type=NO_LOCK.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>