P11SAK: Fix coverity finding: out-of-bounds access
Use a large enough buffer for holding a BER encoded OCET-STRING with a
1 byte length field. This can hold up to 255 bytes.

This is only a theoretical problem, because the call to
EVP_PKEY_get_octet_string_param() to obtain the data returns only up to
buffer size minus 3 bytes, and thus the BER encoded data will never use
a length larger than what fits into the buffer.

Nevertheless, to silence coverity, make the buffer 300 bytes so that
such an out-of-bounds access can not happen, even theoretically.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>