LGR support for PKEY: handle LGR in ep11tok_pkey_update
Make sure that the new pkey's verification pattern matches the one in
ep11_data. This should always be the case, except there was a live
guest relocation (LGR) in the middle of the process. In case of XTS
there could even be an LGR between creating the protected keys from
the two XTS keys. Also add the pkey lock usage at some places where
it was missing in old code.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>