COMMON: Handle CKA_ALWAYS_AUTHENTICATE during operation init
Check attribute CKA_ALWAYS_AUTHENTICATE on operation init calls with private
keys (i.e. C_SignInit, C_DecryptInit) and set a flag in the operation context
that authentication is required before the operation can be continued.

Note that the ICSF token does not support the CKA_ALWAYS_AUTHENTICATE
attribute, because the underlying ICSF services do not support that
attribute. Specifying CKA_ALWAYS_AUTHENTICATE in a template will most
likely fail because ICSF does not know this attribute. The behavior is
as if CKA_ALWAYS_AUTHENTICATE would be FALSE.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>