TRACE: Print the initial trace message for all trace levels
Print the initial trace message as ERROR instead of as INFO, although this
is really not an error. However, when a trace level lower than INFO is
selected, then the initial trace message would not be written.

The initial trace message contains information about the Opencryptoki
version as well as the trace level selected. This information is always
important, even for lower trace levels.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>