P11SAK: Fix clang-tidy finding
The clang-tidy tool reported a suspicious use of sizeof() of a pointer.

sizeof(header) should of course be sizeof(*header), to get the size of
the header structure, and not the size of a pointer to the header.

Luckily it does not make a difference if compiled on a 64 bit architecture,
because the size of the header is 8 bytes, and the size of a pointer is also
8 bytes on a 64 bit system.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>