P11SAK: Fix uninitialized variable warning
Dependent on the object type, get_obj_infos() may or may not set the
keysize, label and common name arguments. Normally, the sort-type should
be set according to the possible object types, but initialize the keysize
variables to silence the warning, and ensure to not call strcmp() with
NULL parameters.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>