COMMON: Add SHAKE key derivation mechanisms to common code
Ensure that the new SHAKE mechanisms are supported by common code
utility functions, as well as policy code.

Add a new token specific SHAKE key derivation function to allow a
token to provide such an implementation. Other than the SHA-based key
derivation functions, SHAKE key derivation can not be implemented by
common code, because there is no SHAKE digest mechanism that could be
used to hash the base key value. Therefore, this must be passed to
the tokens for them to perform the key derivation.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>