TESTCASE: Add SHAKE key derivation tests
Enhance the list of hash based key derivation mechanisms to also
include the two SHAKE key derivation mechanisms.

According to the PKCS#11 standard, SHAKE key derivation needs to
have a key type and/or a length provided in the template. Skip the
tests where either no key type and length is specified at all, or
no length is specified together with a key type of CKK_GENERIC_SECRET.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>