EP11: Add support for PKCS#11 v3.0 SHA3 mechanisms and MGFs for RSA-OAEP
EP11 natively only supports IBM-specific SHA3 mechanisms and mask generation
functions (MGFs) with the RSA-OAEP mechanism, but not the PKCS#11 v3.0 SHA3
mechanisms and MGFs. This affects the 'hashalg' and 'mgf' fields of the
mechanism parameter CK_RSA_PKCS_OAEP_PARAMS of the RSA-OAEP mechanisms.

Transparently translate the PKCS#11 v3.0 SHA3 mechanisms and MGFs to the
corresponding IBM-specific SHA3 mechanisms and MGFs in the mechanism parameter
for RSA-OAEP.

Note that EP11 does not support the SHA3 mechanisms and MGFs for RSA-PSS,
not even the IBM-specific ones. Thus no translation is done for RSA-PSS.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>