EP11: Support SHA3 variants of RSA-PKCS and ECDSA mechanisms
EP11 natively only supports the SHA2 variants of the RSA-PKCS, RSA-PSS
and ECDSA mechanisms.

Transparently split the PKCS#11 v3.0 SHA3 variants of the RSA-PKCS and
ECDSA mechanisms into the hashing part, performed by libica, and the
remaining part performed using the corresponding non-hashed mechanism by
EP11.

These SHA3 variants are only supported when the use of libica is not disabled
in the EP11 token config file.

Note that EP11 does not support the SHA3 mechanisms and MGFs for RSA-PSS,
not even the IBM-specific ones. Furthermore, EP11 requires that the data
hash method matches to the 'hashalg' and 'mgf' fields of the mechanism
parameter CK_RSA_PKCS_PSS_PARAMS for RSA-PSS. Thus, the SHA3 variants of
the RSA-PSS mechanisms are also not possible, because that would require
to also use SHA3 mechanisms and MGFs in the mechanism parameter.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>