SOFT: Correct check for PQC algorithm support
The newest OQS-provider does no longer support the round 2 and 3
dilithium algorithms, only the standardized variants (i.e. ML-DSA).
However, the current check only checked if the algorithm OID is known
by OpenSSL (which is still the case), but not if the algorithm is
really supported.

Add a check to fetch the algorithm to verify if the algorithm is
really supported to avoid failures.

Don't report the CKM_IBM_DILITHIUM mechanism supported if none of the
round 3 variants are supported by the OQS-provider.

Note that OpenCryptoki currently does not yet support the standardized
ML-DSA algorithms, only the round 2 and 3 variants.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>