CCA: Query certain ACPs during token init and APQN event
Find the default role of each CCA APQN used by the token and query
its access control points (ACPs). Certain ACPs are then stored in
the private token data structure so that they can be accessed later.
When more than one APQN is used by the token, then the minimum ACP
setting is stored, i.e. the ACP is only considered to be enabled, if
it is enabled on all APQNs.

Note that if the ACPs change while the token is active, the changes are
not visible for the token. The ACP infos are only queried during token
init, and when an APQN event (offline/online change of an APQN) happens.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>