CONFIGURE: Add --enable-sanitizer option for easy sanitizer builds
Add '--enable-sanitizer' option to easily enable a build with sanitizer.
This requires that libasan and libubsan are available.

When running with sanitizer, disable unloading of the Opencryptoki
shared libraries, to produce useful stack traces.

For the same reason 'export LSAN_OPTIONS=fast_unwind_on_malloc=false'
might be required when running the built executable.
Attention: This may make it very slow!

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>